+++ RELEASE_NOTES.md
@@ -413,4 +413,6 @@
 
 ## Version 1.5.6
 
-- Added File watch function `iot_file_watch' to watch for file and directory changes
+- Added File watch function `iot_file_watch` to watch for file and directory changes
+- Added function `iot_file_exists` to check if a file or directory exists
+- Added functions `iot_file_append_binary` and `iot_file_append` to append data to a file